refactor: MemberUtil 삭제
